+++ CHANGES.md
@@ -3,8 +3,8 @@
 - Better descriptions for rail forks
 - Fix a bug that would cause the rail classifier to go nuts when multiple curved rails shared a tile
 - Fix: only show circuit mode of operation on buffer and requester chests
-
-
+- Remove "add section to start" in the logistics interface. The Factorio API turns out only to support adding sections to the end.
+- Also remove "add section after this" for the same reason.
 # 0.16.21 (2025-11-16)
 
 - Tweak tile sorting. It now goes: